Remove configuration structure
All configuration are managed directly in the Backup class via its member
variables.

+++ src/backup.cpp
@@ -25,22 +25,28 @@ static const std::vector<const char*> networkConfigs = {
 };
 // clang-format on
 
-Backup::Backup(const Configuration& cfg) : config(cfg)
-{
-    std::string tmpPathTemplate = fs::temp_directory_path() / "backup_XXXXXX";
-    tmpDir = mkdtemp(tmpPathTemplate.data());
-}
-
 Backup::~Backup()
 {
-    fs::remove_all(tmpDir);
+    if (!tmpDir.empty())
+    {
+        fs::remove_all(tmpDir);
+    }
 }
 
-void Backup::backup() const
+void Backup::backup()
 {
-    if (config.handleAccounts)
+    if (fs::exists(archiveFile))
     {
-        Accounts acc(config.rootFs, tmpDir, config.readOnlyFs);
+        std::string err = "Backup file already exists: ";
+        err += archiveFile;
+        throw std::runtime_error(err);
+    }
+
+    tmpDir = createTempDir();
+
+    if (handleAccounts)
+    {
+        Accounts acc(rootFs, tmpDir, readOnlyFs);
         acc.backup();
     }
 
@@ -49,7 +55,7 @@ void Backup::backup() const
         backupFile(it);
     }
 
-    if (config.handleNetwork)
+    if (handleNetwork)
     {
         for (const auto& it : networkConfigs)
         {
@@ -57,21 +63,30 @@ void Backup::backup() const
         }
     }
 
-    const Manifest manifest(config.rootFs);
+    const Manifest manifest(rootFs);
     manifest.save(tmpDir);
 
     callTar(false);
 }
 
-void Backup::restore() const
+void Backup::restore()
 {
+    if (!fs::exists(archiveFile))
+    {
+        std::string err = "File not found: ";
+        err += archiveFile;
+        throw std::runtime_error(err);
+    }
+
+    tmpDir = createTempDir();
+
     callTar(true);
 
     checkManifest();
 
-    if (config.handleAccounts)
+    if (handleAccounts)
     {
-        Accounts acc(tmpDir, config.rootFs, config.readOnlyFs);
+        Accounts acc(tmpDir, rootFs, readOnlyFs);
         acc.restore();
     }
 
@@ -80,7 +95,7 @@ void Backup::restore() const
         restoreFile(it);
     }
 
-    if (config.handleNetwork)
+    if (handleNetwork)
     {
         for (const auto& it : networkConfigs)
         {
@@ -89,10 +104,21 @@ void Backup::restore() const
     }
 }
 
+fs::path Backup::createTempDir() const
+{
+    std::string pathTemplate = fs::temp_directory_path() / "backup_XXXXXX";
+    const char* path = mkdtemp(pathTemplate.data());
+    if (!path)
+    {
+        throw std::system_error(errno, std::system_category(), pathTemplate);
+    }
+    return path;
+}
+
 void Backup::checkManifest() const
 {
     const Manifest mnfBackup = Manifest::load(tmpDir);
-    const Manifest mnfCurrent = Manifest(config.rootFs);
+    const Manifest mnfCurrent = Manifest(rootFs);
 
     if (mnfBackup.machineName() != mnfCurrent.machineName())
     {
@@ -103,7 +129,7 @@ void Backup::checkManifest() const
         throw std::runtime_error(err);
     }
 
-    printf("Restore from backup file %s\n", config.backupFile.c_str());
+    printf("Restore from backup file %s\n", archiveFile.c_str());
     mnfBackup.print();
     if (mnfBackup.osVersion() != mnfCurrent.osVersion())
     {
@@ -112,7 +138,7 @@ void Backup::checkManifest() const
                mnfCurrent.osVersion().c_str(), mnfBackup.osVersion().c_str());
     }
 
-    if (!config.unattendedMode)
+    if (!unattendedMode)
     {
         printf("Do you want to continue? [y/N]: ");
         char answer[2]; // first char with \0
@@ -126,11 +152,11 @@ void Backup::checkManifest() const
 
 void Backup::backupFile(const char* path) const
 {
-    fs::path src = config.rootFs / path;
+    fs::path src = rootFs / path;
     if (!fs::exists(src))
     {
         // try to get the file from RO
-        src = config.readOnlyFs / path;
+        src = readOnlyFs / path;
         if (!fs::exists(src))
         {
             return;
@@ -151,14 +177,14 @@ void Backup::restoreFile(const char* path) const
         return;
     }
 
-    const fs::path dst = config.rootFs / path;
+    const fs::path dst = rootFs / path;
 
     // restore permissions
     fs::path permsFile = dst;
     if (!fs::exists(permsFile))
     {
         // try to get permissions from RO
-        permsFile = config.readOnlyFs / path;
+        permsFile = readOnlyFs / path;
     }
     if (fs::exists(permsFile))
     {
@@ -178,7 +204,7 @@ void Backup::callTar(bool extract) const
     std::string cmd = "tar ";
     cmd += extract ? "xf" : "czf";
     cmd += " '";
-    cmd += config.backupFile;
+    cmd += archiveFile;
     cmd += "' -C '";
     cmd += tmpDir;
     cmd += '\'';
@@ -191,7 +217,7 @@ void Backup::callTar(bool extract) const
     if (rc)
     {
         std::string err = "Tar error: ";
-        err += config.backupFile;
+        err += archiveFile;
         throw std::runtime_error(err);
     }
 }
